More secure default server setting
Under some fairly obscure conditions, defaulting the server setting to
'puppet' can be a security concern. Because the worst case scenario
involves a user accidentally running `puppet agent -t` on an untrusted
network, this PR removes the default completely when non-root.
Otherwise, it just prints a deprecation warning.

The check and warnings are generated at the http service so that it
only screams if you actually attempt to use it. This avoids, for example,
the spurious deprecation message when running `puppet config set` to set
this in the first place.

Fixes https://github.com/voxpupuli/security-tracking/issues/22