Introduce the Linux runmode
This builds on the previous work to define all paths in the RunMode
class and has a few interesting aspects.

The parts that make it Linux specific is that it prefers systemd
environment variables to determine locations. When those aren't
available, it instead uses hardcoded paths for root and follows the XDG
specification for non-root.

Currently this mode is not used yet, but by having it in the codebase
distribution packagers can easily use it without having to change all
paths.

Link: https://www.freedesktop.org/software/systemd/man/latest/systemd.exec.html#RuntimeDirectory=
Link: https://specifications.freedesktop.org/basedir-spec/latest/