Perform selabel lookup during file resource synchronization
The default values of the SELinux-related properties of a file resource
are calculated at the beginning of the catalog application phase, before
the catalog resources are actually synchronized. However, by the time
the file resource is synchronized, its default value might be outdated
due to new policy rules added in previous parts of the catalog (for
example, a package was installed that included new policy rules) leading
to the file being created with the wrong SELinux context.

This commit delays the lookup of the default values until the very
moment the file resource is being synchronized. More precisely, the
following process is implemented:

- A `:lookup` symbol is introduced as the default value. It means that
  the actual value is to be looked up later during synchronization.
  However, maintaining the previous behaviour, a default value of `nil`
  is used if either the platform is Windows or the user explicitly
  stated to ignore the default values (using the
  `selinux_ignore_defaults` parameter).
- During resource synchronization, when the `insync?` method is
  executed, the `:lookup` symbol is replaced with the actual looked up
  value and then the sync status is calculated as usual.

The unit test have been adjusted accordingly.