Remove deprecated checksum-in-file-content functionality
It was possible, in the distant past, to set a file's content to a
checksum and have Puppet automagically fetch file content from
filebucket.

This functionality is long deprecated, and was once scheduled for
removal in Puppet 7. I'm not even sure it's working today.

This commit disables the functionality entirely.