Be a bit more verbose when loading module facts
Right now, you only get a rather unhelpful pile of `Loading facts`
messages when having a log level of info. This changes the messages
so that it's clear which module's facts are loaded and you're not
staring at a bunch of info-level messages that are only understandable
when you're actually at debug level.